# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Maintenance Request System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  name?: string;
  department?: string;
  phone?: string;
  isActive?: boolean;
  role?: UserUpdateRole;
}